ChromeBased: Obtain passwords from SecretStorage by schema identifier
This way it does not depend on the browser's name. It also avoids a
surprising ItemNotFoundException that could happen when iterating
through get_all_items().

On Linux, SecretStorage is a dependency of keyring, so we can safely
assume its presence.

Wrap access to SecretStorage with closing(), as recommended by the
documentation.

There was broad exception handling before, which this commit removes.
It might be useful to add back specific exception handling later.